aideveloper828-byte/requirements.txt
openpyxl>=3.1.2
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
requests-oauthlib>=1.3.1
hubspot-api-client>=8.0.0
